Refactor plan limits and add record counting functionality
- Commented out the LimitHit function in planConfig.go to disable limit checks.
- Added CountProjectsByUserID function in projectQueries.go to count projects by user ID.
- Introduced CountRecordsByProjectID function in recordQueries.go to count records for a specific project.
- Enhanced CreateSingleOrMultipleRecords handler in record.go to enforce record limits based on user plans and total values in requests.
- Updated UpdateTable and DeleteTable handlers in table.go to include warnings for destructive changes.
- Modified plans.json to adjust plan limits for projects and records per project.
- Created API documentation in api.md and OpenAPI spec in openapi.yaml for better reference and usage.

+++ internal/handlers/API/table.go
@@ -161,9 +161,11 @@ func UpdateTable(c *gin.Context) {
 		return
 	}
 
+	missingFlagMessage := "Warning, this route will perform destructive changes (refer to the API documentation for more details). Destructive changes are not allowed unless 'allow_destructive' is set to true in the request body"
+
 	allow_destructive := requestBody.AllowDestructive
 	if !allow_destructive {
-		utils.ApiError(c, http.StatusBadRequest, "Warning, this route will perform destructive changes (refer to the API documentation for more details). Destructive changes are not allowed without 'allow_destructive' flag set to true")
+		utils.ApiError(c, http.StatusBadRequest, missingFlagMessage)
 		return
 	}
 
@@ -206,6 +208,22 @@ func DeleteTable(c *gin.Context) {
 
 	// get table name from URL param
 	tableName := c.Param("name")
+
+	// access the json body of the request
+	var requestBody types.SchemaDeleteRequestBody
+
+	missingFlagMessage := "Warning, this route will perform destructive changes (refer to the API documentation for more details). Destructive changes are not allowed unless 'allow_destructive' is set to true in the request body"
+
+	if err := c.BindJSON(&requestBody); err != nil {
+		utils.ApiError(c, http.StatusBadRequest, "invalid request body, "+missingFlagMessage)
+		return
+	}
+
+	allow_destructive := requestBody.AllowDestructive
+	if !allow_destructive {
+		utils.ApiError(c, http.StatusBadRequest, missingFlagMessage)
+		return
+	}
 
 	err := db.DeleteTable(c, project.ID, tableName)
